(bzr r3566.5.11)

+++ launcher/SingleMonitorLauncherIcon.cpp
@@ -26,7 +26,7 @@ namespace unity
 namespace launcher
 {
 
-SingleMonitorLauncherIcon::SingleMonitorLauncherIcon(IconType type, unsigned monitor)
+SingleMonitorLauncherIcon::SingleMonitorLauncherIcon(IconType type, int monitor)
  : SimpleLauncherIcon(type)
  , monitor_(monitor)
 {
@@ -36,10 +36,10 @@ SingleMonitorLauncherIcon::SingleMonitorLauncherIcon(IconType type, unsigned mon
 void SingleMonitorLauncherIcon::UpdateMonitor()
 {
   for (unsigned i = 0; i < monitors::MAX; ++i)
-    SetVisibleOnMonitor(i, i == monitor_);
+    SetVisibleOnMonitor(i, static_cast<int>(i) == monitor_);
 }
 
-void SingleMonitorLauncherIcon::SetMonitor(unsigned monitor)
+void SingleMonitorLauncherIcon::SetMonitor(int monitor)
 {
   if (monitor != monitor_)
   {
@@ -48,7 +48,7 @@ void SingleMonitorLauncherIcon::SetMonitor(unsigned monitor)
   }
 }
 
-unsigned SingleMonitorLauncherIcon::GetMonitor() const
+int SingleMonitorLauncherIcon::GetMonitor() const
 {
   return monitor_;
 }
